GoogleTTSClient/ttsclient/requirements.txt
functions-framework==3.5.0
google-cloud-texttospeech